Refactor-7 moved tool specific utilities to tool-utils and handle all stats in stats-handler
